Code cleaning

+++ DotPacket.Core/DotPacket/Registry/WrongPacketException.cs
@@ -4,7 +4,7 @@ namespace DotPacket.Registry
 {
     public class WrongPacketException : Exception
     {
-        public WrongPacketException(uint id, Type type) : base(
+        public WrongPacketException(byte id, Type type) : base(
             $"Given ID {id} is not the one of the packet '{type.FullName}'"
         )
         {
